JavaScriptをモジュラーアーキテクチャにリファクタリング
一枚岩だった `script.js` ファイルを、より保守しやすく拡張性の高いモジュラーアーキテクチャに分割しました。

主な変更点:
- `main.js`: アプリケーションの全体制御役として、状態管理とイベントリスナーを担当します。
- `ui.js`: 仮想DOMを使用したすべてのDOMレンダリングロジックをカプセル化します。
- `gameService.js`: すべてのゲーム設定をインポートし、ゲームデータへの単一のアクセスポイントを提供します。
- `js/games/`: 各ゲーム固有のロジックと設定を格納する個別のモジュールを配置する新しいディレクトリです。これにより、ゲームの追加や修正が大幅に容易になります。

`index.html` は、新しいエントリーポイントである `js/main.js` を参照するように更新されました。この新しい構造により、関心の分離、保守性、拡張性が向上します。
