+++ index.html
@@ -91,6 +91,6 @@
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script type="module" src="js/script.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
